Add files via upload
The Objective of this Capstone Project is to locate the best location to open a new Bakery Outlet in City New Delhi, India. Using Data Science and Machine Learning Technique, this Project aims to answer this Business problem. With this help of Machine learning, we are able to sort out the location which is well connected to Airport/Bus Station and other public places. The Solution recommends the best neighborhood to open the Bakery.